1.添加数据库事务处理到插入，更新动作中

+++ package/usr_app/src/bestomgw/src/dalitek/src/m1_district.c
@@ -23,6 +23,7 @@ int district_create_handle(payload_t data)
 	int row_number = 0;
 	char time[30];
 	char sql_1[200];
+    char* errorMsg = NULL;
 
 	if(data.pdu == NULL) return M1_PROTOCOL_FAILED;
 	getNowTime(time);
@@ -39,42 +40,50 @@ int district_create_handle(payload_t data)
 	char* sql = "select ID from district_table order by ID desc limit 1";
 	/*linkage_table*/
 	id = sql_id(db, sql);
+    if(sqlite3_exec(db, "BEGIN", NULL, NULL, &errorMsg)==SQLITE_OK){
+        printf("BEGIN\n");
+    	/*获取收到数据包信息*/
+        districtNameJson = cJSON_GetObjectItem(data.pdu, "districtName");
+        printf("districtName:%s\n",districtNameJson->valuestring);
+        apIdArrayJson = cJSON_GetObjectItem(data.pdu, "apId");
+        number1 = cJSON_GetArraySize(apIdArrayJson);
+        printf("number1:%d\n",number1);
 
-	/*获取收到数据包信息*/
-    districtNameJson = cJSON_GetObjectItem(data.pdu, "districtName");
-    printf("districtName:%s\n",districtNameJson->valuestring);
-    apIdArrayJson = cJSON_GetObjectItem(data.pdu, "apId");
-    number1 = cJSON_GetArraySize(apIdArrayJson);
-    printf("number1:%d\n",number1);
+       	/*删除原有表scenario_table中的旧scenario*/
+    	sprintf(sql_1,"select ID from district_table where DIS_NAME = \"%s\";",districtNameJson->valuestring);	
+    	row_number = sql_row_number(db, sql_1);
+    	printf("row_number:%d\n",row_number);
+    	if(row_number > 0){
+    		sprintf(sql_1,"delete from district_table where DIS_NAME = \"%s\";",districtNameJson->valuestring);				
+    		sqlite3_reset(stmt);
+    		sqlite3_prepare_v2(db, sql_1, strlen(sql_1), &stmt, NULL);
+    		while(thread_sqlite3_step(&stmt, db) == SQLITE_ROW);
+    	}
+        
+        /*存取到数据表scenario_table中*/
+        for(i = 0; i < number1; i++){
+    		apIdJson = cJSON_GetArrayItem(apIdArrayJson, i);
+    		printf("apId:%s\n",apIdJson->valuestring);
 
-   	/*删除原有表scenario_table中的旧scenario*/
-	sprintf(sql_1,"select ID from district_table where DIS_NAME = \"%s\";",districtNameJson->valuestring);	
-	row_number = sql_row_number(db, sql_1);
-	printf("row_number:%d\n",row_number);
-	if(row_number > 0){
-		sprintf(sql_1,"delete from district_table where DIS_NAME = \"%s\";",districtNameJson->valuestring);				
-		sqlite3_reset(stmt);
-		sqlite3_prepare_v2(db, sql_1, strlen(sql_1), &stmt, NULL);
-		while(thread_sqlite3_step(&stmt, db) == SQLITE_ROW);
-	}
-    
-    /*存取到数据表scenario_table中*/
-    for(i = 0; i < number1; i++){
-		apIdJson = cJSON_GetArrayItem(apIdArrayJson, i);
-		printf("apId:%s\n",apIdJson->valuestring);
-
-		sql = "insert into district_table(ID, DIS_NAME, AP_ID, TIME) values(?,?,?,?);";
-		printf("sql:%s\n",sql);
-		sqlite3_reset(stmt);
-		sqlite3_prepare_v2(db, sql, strlen(sql), &stmt, NULL);
-		sqlite3_bind_int(stmt, 1, id);
-		id++;
-		sqlite3_bind_text(stmt, 2, districtNameJson->valuestring, -1, NULL);
-		sqlite3_bind_text(stmt, 3, apIdJson->valuestring, -1, NULL);
-		sqlite3_bind_text(stmt, 4, time, -1, NULL);
-		rc = thread_sqlite3_step(&stmt, db); 
-		printf("step() return %s\n", rc == SQLITE_DONE ? "SQLITE_DONE": rc == SQLITE_ROW ? "SQLITE_ROW" : "SQLITE_ERROR"); 
-    }
+    		sql = "insert into district_table(ID, DIS_NAME, AP_ID, TIME) values(?,?,?,?);";
+    		printf("sql:%s\n",sql);
+    		sqlite3_reset(stmt);
+    		sqlite3_prepare_v2(db, sql, strlen(sql), &stmt, NULL);
+    		sqlite3_bind_int(stmt, 1, id);
+    		id++;
+    		sqlite3_bind_text(stmt, 2, districtNameJson->valuestring, -1, NULL);
+    		sqlite3_bind_text(stmt, 3, apIdJson->valuestring, -1, NULL);
+    		sqlite3_bind_text(stmt, 4, time, -1, NULL);
+    		rc = thread_sqlite3_step(&stmt, db); 
+    		printf("step() return %s\n", rc == SQLITE_DONE ? "SQLITE_DONE": rc == SQLITE_ROW ? "SQLITE_ROW" : "SQLITE_ERROR"); 
+        }
+        if(sqlite3_exec(db, "COMMIT", NULL, NULL, &errorMsg) == SQLITE_OK){
+            printf("END\n");
+        }
+    }else{
+        printf("errorMsg:");
+    }    
+    sqlite3_free(errorMsg);
 
     sqlite3_finalize(stmt);
     sqlite3_close(db);
